+++ j2v8-debugger/src/main/java/com/alexii/j2v8debugger/Debugger.kt
@@ -7,9 +7,7 @@ import com.eclipsesource.v8.inspector.V8Inspector
 import com.facebook.stetho.inspector.jsonrpc.JsonRpcPeer
 import com.facebook.stetho.inspector.jsonrpc.JsonRpcResult
 import com.facebook.stetho.inspector.protocol.ChromeDevtoolsMethod
-import com.facebook.stetho.inspector.protocol.module.Runtime.RemoteObject
 import com.facebook.stetho.json.ObjectMapper
-import com.facebook.stetho.json.annotation.JsonProperty
 import com.facebook.stetho.websocket.CloseCodes
 import kotlinx.coroutines.runBlocking
 import org.json.JSONObject
@@ -32,6 +30,7 @@ fun urlToScriptId(url: String?) = url?.removePrefix(scriptsUrlBase)
  *  Otherwise setting breakpoint, etc. makes no effect.
  *  Above is also true for the case when debugger is paused due to pause/resume implementation as thread suspension.
  */
+@Suppress("unused")
 class Debugger(
     private val scriptSourceProvider: ScriptSourceProvider
 ) : FacebookDebuggerStub() {
@@ -247,124 +246,5 @@ class Debugger(
         }
     }
 
-    class EvaluateOnCallFrameResult(
-        @field:JsonProperty
-        @JvmField
-        val result: JSONObject? = null
-    ): JsonRpcResult
-
-    /**
-     * Fired as the result of [Debugger.enable]
-     */
-    class ScriptParsedEvent(
-        @field:JsonProperty @JvmField
-        val scriptId: String?,
-
-        @field:JsonProperty @JvmField
-        val url: String? = scriptIdToUrl(scriptId)
-    )
-
-    class GetScriptSourceRequest : JsonRpcResult {
-        @field:JsonProperty
-        @JvmField
-        var scriptId: String? = null
-    }
-
-    class GetScriptSourceResponse(
-        @field:JsonProperty @JvmField
-        val scriptSource: String
-    ) : JsonRpcResult
-
-    class SetBreakpointByUrlRequest : JsonRpcResult {
-        //script id
-        @field:JsonProperty
-        @JvmField
-        var url: String? = null
-
-        @field:JsonProperty
-        @JvmField
-        var lineNumber: Int? = null
-
-        //unused for now
-        @field:JsonProperty
-        @JvmField
-        var columnNumber: Int? = null
-
-        //unused for now
-        @field:JsonProperty
-        @JvmField
-        var condition: String? = null
-
-        val scriptId get() = urlToScriptId(url)
-    }
-
-
-    class SetBreakpointByUrlResponse(
-        @field:JsonProperty @JvmField
-        val breakpointId: String,
-
-        location: Location
-    ) : JsonRpcResult {
-        @field:JsonProperty
-        @JvmField
-        val locations: List<Location> = listOf(location)
-    }
-
-    class RemoveBreakpointRequest : JsonRpcResult {
-        //script id
-        @field:JsonProperty
-        @JvmField
-        var breakpointId: String? = null
-    }
-
-
-    data class Location(
-        @field:JsonProperty @JvmField
-        val scriptId: String,
-
-        @field:JsonProperty @JvmField
-        val lineNumber: Int,
-
-        @field:JsonProperty @JvmField
-        val columnNumber: Int
-    )
-
-    data class CallFrame @JvmOverloads constructor(
-        @field:JsonProperty @JvmField
-        val callFrameId: String,
-
-        @field:JsonProperty @JvmField
-        val functionName: String,
-
-        @field:JsonProperty @JvmField
-        val location: Location,
-
-        /** JavaScript script name or url. */
-        @field:JsonProperty @JvmField
-        val url: String,
-
-        @field:JsonProperty @JvmField
-        val scopeChain: List<Scope>,
-
-        //xxx: check how and whether it's wotking with this
-        @field:JsonProperty @JvmField
-        val `this`: RemoteObject? = null
-    )
-
-    data class Scope(
-        /** one of: global, local, with, closure, catch, block, script, eval, module. */
-        @field:JsonProperty @JvmField
-        val type: String,
-        /**
-         * Object representing the scope.
-         * For global and with scopes it represents the actual object;
-         * for the rest of the scopes, it is artificial transient object enumerating scope variables as its properties.
-         */
-        @field:JsonProperty @JvmField
-        val `object`: RemoteObject
-    )
 }
 
-class SimpleIntegerResult(@JsonProperty(required = true) var result: Int) : JsonRpcResult {
-
-}
